Delete BmpViewerSDI.ncb
Q:"There is a ncb file under the working directory of my Visual Studio 2005 project. It is very big, almost 7M. I am wondering,"
A:"It's the Intellisense database, with information needed to provide auto-completion and such for your code.

You're free to remove this file at any time. It'll be rebuilt."

Reference
https://social.msdn.microsoft.com/Forums/vstudio/en-US/85a08a1f-b39b-4217-9d3c-40dab3aa000b/visual-studio-ncb-file